Rest client validations

+++ src/main/java/enzoDevoto/apps/medicineTakeCareclient/web/client/MedicineTakeCareClient.java
@@ -25,36 +25,63 @@ public class MedicineTakeCareClient {
         this.restTemplate = restTemplateBuilder.build();
     }
 
-    public PatientDto getPatientById(UUID patientUuid){
-        return restTemplate.getForObject(host+ Utils.PATIENT_PATH + patientUuid, PatientDto.class);
+    public PatientDto getPatientById(UUID patientUuid) {
+        return restTemplate.getForObject(host + Utils.PATIENT_PATH + patientUuid, PatientDto.class);
     }
 
     public ResponseEntity<PatientDto> setNewPatient(PatientDto patientDto) {
 
-        return restTemplate.postForEntity(host.concat(Utils.PATIENT_PATH), patientDto, PatientDto.class);
+        if (patientDto.getId() == null) {
+            throw new NullPointerException("This Patient ID is not correct. Patient doesn't exists.");
+        } else {
+            return restTemplate.postForEntity(host.concat(Utils.PATIENT_PATH), patientDto, PatientDto.class);
+        }
+
 
     }
-    public void updatePatient(UUID patientId, PatientDto patient){
-        restTemplate.put(host+Utils.PATIENT_PATH.concat("updatePatient/")+patientId,patient);
+
+    public void updatePatient(UUID patientId, PatientDto patient) {
+        if (patient.getId() == null) {
+            throw new NullPointerException("This Patient ID is not correct. Patient doesn't exists.");
+        } else {
+            restTemplate.put(host + Utils.PATIENT_PATH.concat("updatePatient/") + patientId, patient);
+        }
     }
 
-    public void deletePatient(UUID patientId){
-        restTemplate.delete(host+Utils.PATIENT_PATH.concat("deletePatient/")+patientId);
+    public void deletePatient(UUID patientId) {
+        if (patientId == null) {
+            throw new NullPointerException("This Patient ID is not correct. Patient doesn't exists.");
+        } else {
+            restTemplate.delete(host + Utils.PATIENT_PATH.concat("deletePatient/") + patientId);
+        }
     }
 
     public DoctorDto getDoctorByID(UUID doctorUuid) {
-        return restTemplate.getForObject(host+Utils.DOCTOR_PATH+doctorUuid, DoctorDto.class);
+        if (doctorUuid == null) {
+            throw new NullPointerException("This Patient ID is not correct. Patient doesn't exists.");
+        } else {
+            return restTemplate.getForObject(host + Utils.DOCTOR_PATH + doctorUuid, DoctorDto.class);
+        }
     }
 
     public ResponseEntity<DoctorDto> setNewDoctor(DoctorDto doctorDto) {
         return restTemplate.postForEntity(host.concat(Utils.DOCTOR_PATH), doctorDto, DoctorDto.class);
 
     }
-    public void updateDoctor(UUID doctorId, DoctorDto doctor){
-        restTemplate.put(host+Utils.DOCTOR_PATH.concat("updateDoctor/")+doctorId,doctor);
+
+    public void updateDoctor(UUID doctorId, DoctorDto doctor) {
+        if (doctorId == null) {
+            throw new NullPointerException("This Patient ID is not correct. Patient doesn't exists.");
+        } else {
+            restTemplate.put(host + Utils.DOCTOR_PATH.concat("updateDoctor/") + doctorId, doctor);
+        }
     }
 
-    public void deleteDoctor(UUID doctorId){
-        restTemplate.delete(host+Utils.DOCTOR_PATH.concat("deleteDoctor/")+doctorId);
+    public void deleteDoctor(UUID doctorId) {
+        if (doctorId == null) {
+            throw new NullPointerException("This Patient ID is not correct. Patient doesn't exists.");
+        } else {
+            restTemplate.delete(host + Utils.DOCTOR_PATH.concat("deleteDoctor/") + doctorId);
+        }
     }
 }
